beets: Chuck music into alphabetical buckets
It like 99% works. I want Various Artists Music folder to be its own top
folder and have the folder name have the year on the end of the name
instead of the begining, but it screwed up the name for all other
albums, so for the time being I'll leave it how it is.

Changed music player to MPV for the time being.