vim: revert to legacy version of ron colorscheme
For the up and coming vim9 release, the devs are putting in a lot of effort to
fix the broken and inconsistent colorschemes, which imho is a good
thing.
But...
The colorscheme I used, "ron" was the most broken, inconsistent and
barren of the lot, so when they fixed it up it had the largest visual
change to what I am use to.

So for the moment, the "temporary" fix is to use the legacy version.
In reality most of my liked colors are getting inherited from the
terminal so I should at some point create a standalone colorscheme