bash: check distro ID only
Instead of checking all the distro's release information, we just check
the distros ID specifically. Avoids the "No LSB modules are available."
being printed out to the standard terminal on systems that only just
have the lsb-release version reporting tool (E.g. nixos with the
pkg.lsb-release installed). At somepoint in the future I should check if
the binary/script exists before trying to use it's output.

Just to note for future reference, running lsb-release without arguments
defults it to using -v (version shorthand argument). Which has the
potential to print "No LSB modules are available." and the only ther
case where that can pop up is when running with -a/--all flag.